refactor: collapsed in one file take order actions and tested use case of take and cancel
